refactor: change and rename app and pages
